Change gamemode to (5,8) and add optionality for custom game settings (#7)
* Change gamemode to (5,8) and add optionality for custom game settings

* Add unit test for invalid game settings

+++ autochess/programs/autochess/src/state/entities.rs
@@ -116,8 +116,8 @@ impl Entities {
         None
     }
 
-    pub fn count_for_controller(&self, controller: Controller) -> u16 {
-        let mut count: u16 = 0;
+    pub fn count_for_controller(&self, controller: Controller) -> u8 {
+        let mut count: u8 = 0;
         for entity in &self.all {
             if entity.owner == controller {
                 count += 1;
